SSCS-10693 refactor to improve layout
